feat: add BrowserRouter
