+++ src/components/app/App.module.css
@@ -3,12 +3,28 @@
   max-width: 1440px;
   margin: 0 auto;
   height: 100%;
-  background-color: #e5e5e5;
 }
 
-.main {
-  display: flex;
+a {
+  color: #3498db;
+  font-size: 25px;
+}
+
+a:hover {
+  color: #1abc9c;
+}
+
+nav {
+  background-color: #2c3e50;
   padding: 10px;
-  align-items: center;
-  flex-direction: column;
+}
+
+nav a {
+  margin: 15px;
+}
+
+nav a.active:hover {
+  color: #ffffff;
+  text-decoration: none;
+  cursor: default;
 }
