+++ src/components/main-content/MainContent.module.css
@@ -1,0 +1,6 @@
+.main {
+  display: flex;
+  padding: 10px;
+  align-items: center;
+  flex-direction: column;
+}
