+++ src/components/main-layout/MainLayout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router-dom';
+import Menu from '../menu/Menu';
+
+export default function MainLayout() {
+  return (
+    <>
+      <Menu />
+      <Outlet />
+    </>
+  );
+}
