+++ src/components/menu/Menu.tsx
@@ -1,0 +1,12 @@
+import { NavLink } from 'react-router-dom';
+
+export default function Menu() {
+  return (
+    <nav>
+      <NavLink to="." end>
+        Home
+      </NavLink>
+      <NavLink to="about">About</NavLink>
+    </nav>
+  );
+}
